rename `hostedServices` -> `gitHost`
