+++ app/src/lib/branches/service.ts
@@ -3,8 +3,8 @@ import { observableToStore, storeToObservable } from '$lib/rxjs/store';
 import { buildContextStore } from '$lib/utils/context';
 import { Observable, combineLatest, of } from 'rxjs';
 import { catchError, shareReplay, startWith, switchMap } from 'rxjs/operators';
-import type { HostedGitListingService } from '$lib/hostedServices/interface/hostedGitListingService';
-import type { PullRequest } from '$lib/hostedServices/interface/types';
+import type { HostedGitListingService } from '$lib/gitHost/interface/hostedGitListingService';
+import type { PullRequest } from '$lib/gitHost/interface/types';
 import type { RemoteBranchService } from '$lib/stores/remoteBranches';
 import type { Branch, RemoteBranch } from '$lib/vbranches/types';
 import type { VirtualBranchService } from '$lib/vbranches/virtualBranch';
